add theme and font option tests to PreferenceTest
